requirements.txt
fastapi
uvicorn
python-dotenv
requests
pinecone
pydantic
huggingface_hub
numpy
groq
